# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceCalibrationOut": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "deviceFingerprint": {
            "title": "Devicefingerprint",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "offsetMs": {
            "title": "Offsetms",
            "type": "number"
          },
          "outputLabel": {
            "title": "Outputlabel",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "deviceFingerprint",
          "outputLabel",
          "offsetMs",
          "createdAt",
          "updatedAt"
        ],
        "title": "DeviceCalibrationOut",
        "type": "object"
      },
      "DeviceCalibrationWrite": {
        "additionalProperties": false,
        "properties": {
          "deviceFingerprint": {
            "minLength": 1,
            "title": "Devicefingerprint",
            "type": "string"
          },
          "offsetMs": {
            "maximum": 1000.0,
            "minimum": -1000.0,
            "title": "Offsetms",
            "type": "number"
          },
          "outputLabel": {
            "minLength": 1,
            "title": "Outputlabel",
            "type": "string"
          }
        },
        "required": [
          "deviceFingerprint",
          "outputLabel",
          "offsetMs"
        ],
        "title": "DeviceCalibrationWrite",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API and WebSocket clock/transport gateway. Clients expand tempo-map beats locally; the server only agrees on revision, anchor, and absolute server time.",
    "title": "FeelMyRythm API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/calibrations": {
      "get": {
        "operationId": "list_calibrations_api_calibrations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceCalibrationOut"
                  },
                  "title": "Response List Calibrations Api Calibrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Calibrations",
        "tags": [
          "device calibrations"
        ]
      },
      "put": {
        "operationId": "upsert_calibration_api_calibrations_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCalibrationWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCalibrationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Calibration",
        "tags": [
          "device calibrations"
        ]
      }
    },
    "/api/calibrations/{calibration_id}": {
      "delete": {
        "operationId": "delete_calibration_api_calibrations__calibration_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "calibration_id",
            "required": true,
            "schema": {
              "title": "Calibration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Calibration",
        "tags": [
          "device calibrations"
        ]
      },
      "get": {
        "operationId": "get_calibration_api_calibrations__calibration_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "calibration_id",
            "required": true,
            "schema": {
              "title": "Calibration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCalibrationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Calibration",
        "tags": [
          "device calibrations"
        ]
      }
    },
...
  }
}